backend/requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
google-auth==2.28.2
PyJWT==2.11.0
pymysql==1.1.1
python-dotenv==1.2.1
requests==2.32.3
sqlparse==0.5.5
tzdata==2025.3
